+++ Temperature.sql
@@ -1,7 +1,7 @@
 create table dbo.Temperature(
 	Id int not null Identity(1,1),
-	Temperature int null,
-	Humidity int null,
+	Temperature decimal(9,2) null,
+	Humidity decimal(9,2)  null,
 	Date Datetime null,
 	constraint PK_dbo_Temperature primary key (Id)
 );
